Add Retina Comment Back

+++ cpp/pixels-retina/include/RGVisibility.h
@@ -30,7 +30,7 @@ public:
     explicit RGVisibility(uint64_t rgRecordNum, uint64_t timestamp, const std::vector<uint64_t>& initialBitmap);
     ~RGVisibility() override;
 
-    void deleteRGRecord(uint16_t rowId, uint64_t timestamp);
+    void deleteRGRecord(uint32_t rowId, uint64_t timestamp);
     uint64_t* getRGVisibilityBitmap(uint64_t timestamp);
 
     void collectRGGarbage(uint64_t timestamp);
@@ -38,15 +38,15 @@ public:
     uint64_t getBitmapSize() const;
 
 private:
-    // Now derived from template parameter
     static constexpr uint32_t VISIBILITY_RECORD_CAPACITY = CAPACITY;
     static constexpr uint32_t BITMAP_SIZE_PER_TILE_VISIBILITY = BITMAP_WORDS(CAPACITY);
 
     TileVisibility<CAPACITY>* tileVisibilities;
     const uint64_t tileCount;
 
-    TileVisibility<CAPACITY>* getTileVisibility(uint16_t rowId) const;
+    TileVisibility<CAPACITY>* getTileVisibility(uint32_t rowId) const;
 };
+
 static_assert(RETINA_CAPACITY > 0 && RETINA_CAPACITY % 64 == 0,
               "RETINA_CAPACITY must be a multiple of 64.");
 using RGVisibilityInstance = RGVisibility<RETINA_CAPACITY>;
